Set up server.js
Using NodeJS and SocketIO, we create a server and pass certain metadata to it. The purpose of this is so that we can communicate our user's information, decisions, and touch inputs with the game.

We create a net server and assign a port number that the game is listening on for specific events. These events are then interpreted and handled by the game.