Create index.html
This is the webpage that gets sent through the web socket. The users are presented with a form that asks for their name and email. They are then presented with options; they can choose to play on the installation, or control it using the tablet. The tablet emit particular events which get sent to the server and interpreted by the game. These include things such as player name, finger position and id, number of fingers, and the color they selected from the final screen. When the user is finished playing, the game will send the user a link where they can claim, download, and share their song.